size_t is the result from the sizeof operation and is not the largest value type, the largest value type is uintmax_t.

+++ include/seagrass/error.h
@@ -5,6 +5,6 @@
 
 #define SEAGRASS_ERROR_NONE                                 0
 
-extern _Thread_local size_t seagrass_error;
+extern _Thread_local uintmax_t seagrass_error;
 
 #endif /* _SEAGRASS_ERROR_H_ */
