+++ src/error.c
@@ -1,4 +1,4 @@
 #include <stdlib.h>
 #include <seagrass.h>
 
-_Thread_local size_t seagrass_error = SEAGRASS_ERROR_NONE;
+_Thread_local uintmax_t seagrass_error = SEAGRASS_ERROR_NONE;
